original-code/requirements.txt
numpy==1.26.4
scikit-learn~=1.6.1
python-dotenv==1.0.1
pandas==2.2.3
torch==2.6.0
transformers==4.50.3
tqdm==4.67.1